Refactor use of AsyncStorage into StorageManager
