+++ app/components/StatusCard.js
@@ -1,7 +1,6 @@
 import React, {Component} from 'react';
 import { StyleSheet, View, Image, Button, TouchableOpacity } from 'react-native';
 import { Text, Divider } from 'react-native-elements';
-import AsyncStorage from '@react-native-async-storage/async-storage';
 import FontAwesome5 from 'react-native-vector-icons/FontAwesome5';
 
 import Moment from 'moment';
